Upgrade to Yarn v4 for local development
Hit an issue on SiteSwan in CI where it tried to install Yarn classic to
pack the dependency since we point to a fork/branch, but then the
checksum of the packed archive didn't match what the lockfile called
for.

This seemed to pop up out of nowhere. I noticed we used to install Yarn
v1.22.19 for this step, but now it's installing Yarn v1.22.22, so maybe
that's a culprit?

Ultimately switching our lockfile to the newer Yarn avoids SiteSwan CI
installing Classic Yarn for packing, which seems to resolve our issue.